Add caching to gtts

+++ src/tts_gtts.py
@@ -25,7 +25,7 @@ def find_player():
 def play(filename: str):
     external_player = find_player()
     if external_player:
-        subprocess.call([external_player, filename])
+        subprocess.run([external_player, filename])
     else:
         import pygame
         pygame.init()
@@ -38,16 +38,33 @@ def play(filename: str):
         pygame.quit()
 
 class TTS:
-    def __init__(self, lang='en'):
+    def __init__(self, lang='en', block=True):
         self.lang = lang
-        self.workdir = tempfile.mkdtemp()
-        self.workfile = os.path.join(self.workdir, "speech.mp3")
+        self.workdir = os.path.join(os.getcwd(), 'gtts')
+        
+        if not os.path.exists(self.workdir):
+            os.makedirs(self.workdir)
 
-    def speak(self, text):
-        stream = BytesIO()
+    def speak(self, text, block):
+        should_be_cached = False
+        
+        if len (text) < 80:
+            should_be_cached = True
+            name = text
+        else:
+            name = text[:61] + "..."
+            
+        self.workfile = os.path.join(self.workdir, f"{name}.mp3")
+        
+        if os.path.exists(self.workfile):
+            play(self.workfile)
+            return True
+        
         tts = gTTS(text, lang=self.lang)
-        # tts.write_to_fp(stream)
         tts.save(self.workfile)
         play(self.workfile)
-        os.unlink(self.workfile)
+    
+        if not should_be_cached:
+            os.unlink(self.workfile)
+    
         return True
